refactored component logic to the redux reducer
